+++ packages/components/src/components/buttons/TooltipButton/TooltipButton.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import styled from 'styled-components';
 import { Icon } from '../../Icon';
 import { Tooltip } from '../../Tooltip';
-import { Button, ButtonProps } from '../Button';
+import { Button, ButtonProps } from '../Button/Button';
 import { useTheme } from '../../../utils';
 
 const StyledButton = styled(Button)`
