+++ packages/components/src/components/others/SecurityCard/index.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import styled from 'styled-components';
 import { CardProps, Card } from '../Card';
 import { IconProps, Icon } from '../../Icon';
-import { Button } from '../../buttons/Button';
+import { Button } from '../../buttons/Button/Button';
 import { useTheme } from '../../../utils/hooks';
 import * as variables from '../../../config/variables';
